Created vault-k8s-helm.sh script for local demo.
Instant 3xConsul + 3xVault cluster on Kubernetes local (single-node or more).  The kube and helm parts work on existing kube clusters, but this starts with a local single-node Kube cluster setup via kubeadm.  It was built and tested on Fedora 29 with kubernetes repo enabled and kubeadm+kube v1.14.  Consul helm chart v0.7.0, Vault from incubator with UI enabled and exposed locally on a NodePort.