Redefines the local setup experience for raft
Changes `local-test` to `local`

This is for setting up things locally and it the suffix did not feel
like it added much value to the name

Changes `raft-cluster-demo.sh` to `cluster.sh`

This change creates a more modular approach to the way to do the setup
of the network, config, and each vault in the cluster. It also provides
an alias ability to make it possible to talk to different Vaults without
requiring a new terminal.