Add OpenLDAP secrets engine demonstration
The goal here is to break down the original OpenLDAP + sshd PoC
into a stepwise guide that can serve as a deeper dive bonus round
to the main OpenLDAP secrets engine guide that is currently in
review in https://github.com/hashicorp/learn/pull/1165.

This extends the original PoC to include a Vault container and
work through an example of credential rotation and read.

It will be linked from the main guide as an optional continuation.