Fix the setting of the VAULT_SERVICE env var in vault-raft user data
Based on other uses of the VAULT_SERVICE environment variable in this
repository, it appears that this was originally a typo in the vault
server userdata script. This changes fixes this typo so that the Vault
systemd service is not turned into an empty file when the value of
VAULT_SERVICE is used later in the script.